WIP move WindowCreate and WindowFocusOrCreate into WindowManager
