+++ src/openrct2-ui/interface/Window.h
@@ -55,44 +55,6 @@ namespace OpenRCT2
 
 namespace OpenRCT2::Ui::Windows
 {
-    WindowBase* WindowCreate(
-        std::unique_ptr<WindowBase>&& w, WindowClass cls, ScreenCoordsXY pos, int32_t width, int32_t height, uint32_t flags);
-    template<typename T, typename... TArgs, typename std::enable_if<std::is_base_of<WindowBase, T>::value>::type* = nullptr>
-    T* WindowCreate(
-        WindowClass cls, const ScreenCoordsXY& pos = {}, int32_t width = 0, int32_t height = 0, uint32_t flags = 0,
-        TArgs&&... args)
-    {
-        return static_cast<T*>(WindowCreate(std::make_unique<T>(std::forward<TArgs>(args)...), cls, pos, width, height, flags));
-    }
-    template<typename T, typename... TArgs, typename std::enable_if<std::is_base_of<WindowBase, T>::value>::type* = nullptr>
-    T* WindowCreate(WindowClass cls, int32_t width, int32_t height, uint32_t flags, TArgs&&... args)
-    {
-        return static_cast<T*>(
-            WindowCreate(std::make_unique<T>(std::forward<TArgs>(args)...), cls, {}, width, height, flags | WF_AUTO_POSITION));
-    }
-    template<typename T, typename std::enable_if<std::is_base_of<WindowBase, T>::value>::type* = nullptr>
-    T* WindowFocusOrCreate(WindowClass cls, const ScreenCoordsXY& pos, int32_t width, int32_t height, uint32_t flags = 0)
-    {
-        auto* windowMgr = GetContext()->GetUiContext()->GetWindowManager();
-        auto* w = windowMgr->BringToFrontByClass(cls);
-        if (w == nullptr)
-        {
-            w = WindowCreate<T>(cls, pos, width, height, flags);
-        }
-        return static_cast<T*>(w);
-    }
-    template<typename T, typename std::enable_if<std::is_base_of<WindowBase, T>::value>::type* = nullptr>
-    T* WindowFocusOrCreate(WindowClass cls, int32_t width, int32_t height, uint32_t flags = 0)
-    {
-        auto* windowMgr = GetContext()->GetUiContext()->GetWindowManager();
-        auto* w = windowMgr->BringToFrontByClass(cls);
-        if (w == nullptr)
-        {
-            w = WindowCreate<T>(cls, width, height, flags);
-        }
-        return static_cast<T*>(w);
-    }
-
     void RideConstructionToolupdateEntranceExit(const ScreenCoordsXY& screenCoords);
     void RideConstructionToolupdateConstruct(const ScreenCoordsXY& screenCoords);
     void RideConstructionTooldownConstruct(const ScreenCoordsXY& screenCoords);
