+++ src/openrct2-ui/windows/Changelog.cpp
@@ -21,6 +21,7 @@
 #include <openrct2/localisation/Formatting.h>
 #include <openrct2/platform/Platform.h>
 #include <openrct2/ui/UiContext.h>
+#include <openrct2/ui/WindowManager.h>
 #include <vector>
 
 namespace OpenRCT2::Ui::Windows
@@ -318,7 +319,8 @@ namespace OpenRCT2::Ui::Windows
 
     WindowBase* ChangelogOpen(int personality)
     {
-        auto* window = WindowBringToFrontByClass(WindowClass::Changelog);
+        auto* windowMgr = GetContext()->GetUiContext()->GetWindowManager();
+        auto* window = windowMgr->BringToFrontByClass(WindowClass::Changelog);
         if (window == nullptr)
         {
             // Create a new centred window
@@ -328,7 +330,7 @@ namespace OpenRCT2::Ui::Windows
             int32_t height = (screenHeight * 4) / 5;
 
             auto pos = ChangelogWindow::GetCentrePositionForNewWindow(width, height);
-            auto* newWindow = WindowCreate<ChangelogWindow>(WindowClass::Changelog, pos, width, height, WF_RESIZABLE);
+            auto* newWindow = windowMgr->Create<ChangelogWindow>(WindowClass::Changelog, pos, width, height, WF_RESIZABLE);
             newWindow->SetPersonality(personality);
             return newWindow;
         }
