+++ src/openrct2-ui/windows/EditorScenarioOptions.cpp
@@ -26,6 +26,8 @@
 #include <openrct2/localisation/Formatter.h>
 #include <openrct2/management/Finance.h>
 #include <openrct2/sprites.h>
+#include <openrct2/ui/UiContext.h>
+#include <openrct2/ui/WindowManager.h>
 #include <openrct2/world/Climate.h>
 #include <openrct2/world/Park.h>
 
@@ -1254,6 +1256,8 @@ namespace OpenRCT2::Ui::Windows
 
     WindowBase* EditorScenarioOptionsOpen()
     {
-        return WindowFocusOrCreate<EditorScenarioOptionsWindow>(WindowClass::EditorScenarioOptions, 280, 148, WF_NO_SCROLLING);
+        auto* windowMgr = GetContext()->GetUiContext()->GetWindowManager();
+        return windowMgr->FocusOrCreate<EditorScenarioOptionsWindow>(
+            WindowClass::EditorScenarioOptions, 280, 148, WF_NO_SCROLLING);
     }
 } // namespace OpenRCT2::Ui::Windows
