+++ src/openrct2-ui/windows/NewsOptions.cpp
@@ -282,6 +282,7 @@ namespace OpenRCT2::Ui::Windows
 
     WindowBase* NewsOptionsOpen()
     {
-        return WindowFocusOrCreate<NewsOptionsWindow>(WindowClass::NotificationOptions, WW, WH, WF_CENTRE_SCREEN);
+        auto* windowMgr = GetContext()->GetUiContext()->GetWindowManager();
+        return windowMgr->FocusOrCreate<NewsOptionsWindow>(WindowClass::NotificationOptions, WW, WH, WF_CENTRE_SCREEN);
     }
 } // namespace OpenRCT2::Ui::Windows
