+++ src/openrct2-ui/windows/ServerStart.cpp
@@ -283,7 +283,8 @@ namespace OpenRCT2::Ui::Windows
 
     WindowBase* ServerStartOpen()
     {
-        return WindowFocusOrCreate<ServerStartWindow>(WindowClass::ServerStart, WW, WH, WF_CENTRE_SCREEN);
+        auto* windowMgr = GetContext()->GetUiContext()->GetWindowManager();
+        return windowMgr->FocusOrCreate<ServerStartWindow>(WindowClass::ServerStart, WW, WH, WF_CENTRE_SCREEN);
     }
 } // namespace OpenRCT2::Ui::Windows
